Yet another round of refactoring. This time around was able to remove one of the interfaces for callback behavior which simplified things a little bit.
I've also adopted a pattern by which we have "controller" objects that manage the state around a request/response/retry/etc cycle. Sucks a bit to have mutable state but I haven't been able to find a way around it and at least with the controllers, the state is managed by a single class that is tied to an understanding of how the request is managed.

The RequestDetailProvider interface was supposed to help in making the RequestSender have more responsibility and be the point of understanding how to get a target location based upon if the caller says they are retrying a request or just launching it normally. However, the retry semantics with multi actions and scanner behavior are making the abstraction look inadequate so it's likely on the way out but for now, it holds place.