Beginning refactor of the multi action to work primarily with sets of indexes that we will then use as pointers into the master actions list when building up multiaction requests. Doing this should allow us to hold the state of which indexes go with which request which we can then refer back to on a local error where we need to do a retry, but don't have the response from the server to guide us to the indexes that need a retry. Thinking this could potentially also aid in capturing which host an error occurred on when dealing with errors/retries.
The incarnation of this work as it stands right now is also pushing me to consider changing the sender to take HostAndPort objects directly rather than HRegionLocation objects. Using HRegionLocation objects to group things by host is really a hack if you dig into it. The location is more than just the server, but it's compareTo method only compares on the host which is why it works as a host key in a map. Gross. We can be more explicit with a HostAndPort and clear some of that mess up.